mops/pyproject.toml
[tool.poetry]
name = "mops"
version = "0.2.2"
description = "Create MOPs and CDs"
authors = ["Josh Dickman"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
keyring = "^23.2.1"
argparse = "^1.4.0"
jinja2 = "^3.0.1"
atlassian-python-api = "^3.14.0"
google-api-core = "^2.0.1"
google-api-python-client = "^2.23.0"
google-auth = "^2.2.1"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
googleapis-common-protos = "^1.53.0"
pyotp = "^2.6.0"
cerebrus = "^0.2"
Cerberus = "^1.3.4"
validators = "^0.18.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
